+++ bofire/any/objective.py
@@ -1,30 +1,30 @@
 from typing import Union
 
-from bofire.domain import objectives
+from bofire.domain import objective
 
 AnyObjective = Union[
-    objectives.MaximizeObjective,
-    objectives.MinimizeObjective,
-    objectives.DeltaObjective,
-    objectives.SigmoidObjective,
-    objectives.MaximizeSigmoidObjective,
-    objectives.MinimizeSigmoidObjective,
-    objectives.ConstantObjective,
-    objectives.CloseToTargetObjective,
-    objectives.TargetObjective,
+    objective.MaximizeObjective,
+    objective.MinimizeObjective,
+    objective.DeltaObjective,
+    objective.SigmoidObjective,
+    objective.MaximizeSigmoidObjective,
+    objective.MinimizeSigmoidObjective,
+    objective.ConstantObjective,
+    objective.CloseToTargetObjective,
+    objective.TargetObjective,
 ]
 
 AnyAbstractObjective = Union[
-    objectives.IdentityObjective,
-    objectives.MaximizeObjective,
-    objectives.MinimizeObjective,
-    objectives.DeltaObjective,
-    objectives.SigmoidObjective,
-    objectives.MaximizeSigmoidObjective,
-    objectives.MinimizeSigmoidObjective,
-    objectives.ConstantObjective,
-    objectives.AbstractTargetObjective,
-    objectives.CloseToTargetObjective,
-    objectives.TargetObjective,
-    objectives.BotorchConstrainedObjective,
+    objective.IdentityObjective,
+    objective.MaximizeObjective,
+    objective.MinimizeObjective,
+    objective.DeltaObjective,
+    objective.SigmoidObjective,
+    objective.MaximizeSigmoidObjective,
+    objective.MinimizeSigmoidObjective,
+    objective.ConstantObjective,
+    objective.AbstractTargetObjective,
+    objective.CloseToTargetObjective,
+    objective.TargetObjective,
+    objective.BotorchConstrainedObjective,
 ]
