--- bofire/domain/__init__.py
@@ -1,3 +1,0 @@
-from .domain import Domain
-
-__all__ = ["Domain"]
