+++ bofire/samplers.py
@@ -7,7 +7,6 @@ import torch
 from botorch.utils.sampling import get_polytope_samples
 from pydantic import validate_arguments, validator
 
-from bofire.domain import Domain
 from bofire.domain.constraint import (
     Constraint,
     LinearEqualityConstraint,
@@ -15,6 +14,7 @@ from bofire.domain.constraint import (
     NChooseKConstraint,
     NonlinearInequalityConstraint,
 )
+from bofire.domain.domain import Domain
 from bofire.domain.feature import (
     CategoricalDescriptorInput,
     CategoricalInput,
