+++ bofire/strategies/botorch/sobo.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel, PositiveFloat, validate_arguments, validator
 
 from bofire.domain.constraints import Constraint
 from bofire.domain.feature import Feature
-from bofire.domain.objectives import Objective
+from bofire.domain.objective import Objective
 from bofire.strategies.botorch.base import BotorchBasicBoStrategy
 from bofire.strategies.botorch.utils.objectives import (
     AdditiveObjective,
