+++ bofire/strategies/strategy.py
@@ -9,7 +9,7 @@ from pydantic.types import NonNegativeInt, PositiveInt
 from bofire.domain.constraints import Constraint
 from bofire.domain.domain import Domain
 from bofire.domain.feature import Feature, OutputFeature, TInputTransformSpecs
-from bofire.domain.objectives import Objective
+from bofire.domain.objective import Objective
 
 
 def validate_constraints(cls, domain: Domain):
