+++ bofire/utils/torch_tools.py
@@ -8,14 +8,14 @@ from torch import Tensor
 from torch.nn import Module
 from torch.nn.functional import one_hot
 
-from bofire.domain import Domain
 from bofire.domain.constraint import (
     LinearEqualityConstraint,
     LinearInequalityConstraint,
 )
+from bofire.domain.domain import Domain
 from bofire.domain.feature import InputFeature
 from bofire.domain.features import OutputFeatures
-from bofire.domain.objectives import BotorchConstrainedObjective
+from bofire.domain.objective import BotorchConstrainedObjective
 
 tkwargs = {
     "dtype": torch.double,
