+++ tests/bofire/strategies/botorch/test_qehvi.py
@@ -13,7 +13,7 @@ from botorch.acquisition.multi_objective.objective import WeightedMCMultiOutputO
 from bofire.benchmarks.multi import C2DTLZ2, DTLZ2
 from bofire.domain.domain import Domain
 from bofire.domain.feature import ContinuousInput, ContinuousOutput
-from bofire.domain.objectives import MaximizeObjective, MinimizeObjective
+from bofire.domain.objective import MaximizeObjective, MinimizeObjective
 from bofire.samplers import PolytopeSampler
 from bofire.strategies.botorch.qehvi import BoTorchQehviStrategy, BoTorchQnehviStrategy
 from bofire.utils.enum import CategoricalMethodEnum
