+++ tests/bofire/test_samplers.py
@@ -1,11 +1,11 @@
 import pytest
 
-from bofire.domain import Domain
 from bofire.domain.constraint import (
     LinearEqualityConstraint,
     LinearInequalityConstraint,
 )
 from bofire.domain.constraints import Constraints
+from bofire.domain.domain import Domain
 from bofire.domain.feature import CategoricalInput, ContinuousInput
 from bofire.domain.features import InputFeatures
 from bofire.samplers import PolytopeSampler, RejectionSampler
